requirements.txt
weaviate-client==4.11.2
numpy==1.26.4
panda==0.3.1
sentence-transformers==3.4.1
fastapi[standard]